+++ apps/payments/create-react-app/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @crossmint/client-sdk-react-ui-starter
 
+## 1.2.30
+
+### Patch Changes
+
+- Updated dependencies [eece432]
+- Updated dependencies [eff943c]
+  - @crossmint/client-sdk-react-ui@1.5.0
+
+## 1.2.29
+
+### Patch Changes
+
+- Updated dependencies [1ae2a5a]
+  - @crossmint/client-sdk-react-ui@1.4.1
+
 ## 1.2.28
 
 ### Patch Changes